// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
      collectors: {
        Row: {
          bio: string | null
          created_at: string | null
          created_by_admin_id: string | null
          email: string | null
          id: string
          location: string | null
          name: string
          notes: string | null
          updated_at: string | null
        }
        Insert: {
          bio?: string | null
          created_at?: string | null
          created_by_admin_id?: string | null
          email?: string | null
          id?: string
          location?: string | null
          name: string
          notes?: string | null
          updated_at?: string | null
        }
        Update: {
          bio?: string | null
          created_at?: string | null
          created_by_admin_id?: string | null
          email?: string | null
          id?: string
          location?: string | null
          name?: string
          notes?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "collectors_created_by_admin_id_fkey"
            columns: ["created_by_admin_id"]
            isOneToOne: false
            referencedRelation: "user_profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      palindromes: {
        Row: {
          additional_notes: string | null
          car_type: string | null
          collector_id: string
          created_at: string | null
          date_found: string | null
          id: string
          image_storage_path: string
          image_url: string
          license_plate: string
          location_found: string | null
          updated_at: string | null
          uploaded_by_admin_id: string
        }
        Insert: {
          additional_notes?: string | null
          car_type?: string | null
          collector_id: string
          created_at?: string | null
          date_found?: string | null
          id?: string
          image_storage_path: string
          image_url: string
          license_plate: string
          location_found?: string | null
          updated_at?: string | null
          uploaded_by_admin_id: string
        }
        Update: {
          additional_notes?: string | null
          car_type?: string | null
          collector_id?: string
          created_at?: string | null
          date_found?: string | null
          id?: string
          image_storage_path?: string
          image_url?: string
          license_plate?: string
          location_found?: string | null
          updated_at?: string | null
          uploaded_by_admin_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "palindromes_collector_id_fkey"
            columns: ["collector_id"]
            isOneToOne: false
            referencedRelation: "collectors"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "palindromes_uploaded_by_admin_id_fkey"
            columns: ["uploaded_by_admin_id"]
            isOneToOne: false
            referencedRelation: "user_profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      user_profiles: {
        Row: {
          avatar_url: string | null
          created_at: string | null
          email: string
          id: string
          is_admin: boolean | null
          name: string | null
          nextauth_user_id: string | null
          updated_at: string | null
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string | null
          email: string
          id?: string
          is_admin?: boolean | null
          name?: string | null
          nextauth_user_id?: string | null
          updated_at?: string | null
        }
        Update: {
          avatar_url?: string | null
          created_at?: string | null
          email?: string
          id?: string
          is_admin?: boolean | null
          name?: string | null
          nextauth_user_id?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never